drivers: soc: samsung: fix build errors
../../../../../../kernel/samsung/exynos9820/drivers/soc/samsung/debug/exynos-ehld.c:102:23: error: cast to smaller integer type 'int' from 'struct perf_event *' [-Werror,-Wpointer-to-int-cast]
                                                        __func__, cpu, (int)event);
                                                                       ^~~~~~~~~~
../../../../../../kernel/samsung/exynos9820/drivers/soc/samsung/debug/exynos-ehld.c:33:51: note: expanded from macro 'ehld_printk'
                                                  ^~~
1 error generated.

+++ drivers/soc/samsung/debug/exynos-ehld.c
@@ -99,7 +99,7 @@ static int exynos_ehld_start_cpu(unsigned int cpu)
 							 exynos_ehld_callback, NULL);
 		if (IS_ERR(event)) {
 			ehld_printk(0, "@%s: cpu%d event make failed err:%d\n",
-							__func__, cpu, (int)event);
+							__func__, cpu, (long)event);
 			return PTR_ERR(event);
 		} else {
 			ehld_printk(0, "@%s: cpu%d event make success\n", __func__, cpu);
